LayoutTests/fast/dom/StyleSheet/detached-stylesheet-without-wrapper.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css" type="text/css" title="js-test-style">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Test that CSS stylesheets lacking wrappers are pruned from detached CSSOM subtrees. " + 
"This is surprising behavior, but it's a very rare situation, and the behavior actually matches " +
"what both WebKit and Gecko do for DOM nodes in detached subtrees.");

rule = getMatchedCSSRules(document.getElementsByTagName("div")[0]).item(0);
document.documentElement.removeChild(document.head);
shouldBe("rule.parentStyleSheet", "null");

var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
