Safari-609.1.11.1/LayoutTests/fast/dom/StyleSheet/detached-stylesheet-without-wrapper-expected.txt

Test that CSS stylesheets lacking wrappers are pruned from detached CSSOM subtrees. This is surprising behavior, but it's a very rare situation, and the behavior actually matches what both WebKit and Gecko do for DOM nodes in detached subtrees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rule.parentStyleSheet is null
PASS successfullyParsed is true

TEST COMPLETE
